Unsubscribing from events OnDestroy

+++ Assets/_UI/Menu/ShowSavedData.cs
@@ -26,29 +26,31 @@ namespace Randolph.Levels {
         }
 
         void DisplayPlayerPrefs() {
-            Vector2 rectSize = new Vector2(200, 30);
-            CurrentOffset = 10;
+            if (Application.isEditor) {
+                var rectSize = new Vector2(200, 30);
+                CurrentOffset = 10;
 
-            int levelIndex = PlayerPrefs.HasKey(levelKey) ? PlayerPrefs.GetInt(levelKey) : 0;
-            GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
-                    $"<color=red>{nameof(levelKey)}</color>: <color=blue>{levelIndex}</color>");
+                int levelIndex = PlayerPrefs.HasKey(levelKey) ? PlayerPrefs.GetInt(levelKey) : 0;
+                GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
+                        $"<color=red>{nameof(levelKey)}</color>: <color=blue>{levelIndex}</color>");
 
-            int checkpointIndex = PlayerPrefs.HasKey(checkpointKey) ? PlayerPrefs.GetInt(checkpointKey) : 0;
-            GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
-                    $"<color=red>{nameof(checkpointKey)}</color>: <color=blue>{checkpointIndex}</color>");
+                int checkpointIndex = PlayerPrefs.HasKey(checkpointKey) ? PlayerPrefs.GetInt(checkpointKey) : 0;
+                GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
+                        $"<color=red>{nameof(checkpointKey)}</color>: <color=blue>{checkpointIndex}</color>");
 
-            GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
-                    $"<color=red>{nameof(LanguageSwitch.LanguageKey)}</color>: <color=blue>{LanguageSwitch.Language}</color>");
+                GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
+                        $"<color=red>{nameof(LanguageSwitch.LanguageKey)}</color>: <color=blue>{LanguageSwitch.Language}</color>");
 
-            bool mute = (PlayerPrefs.HasKey(muteKey) && PlayerPrefs.GetInt(muteKey) == 1);
-            GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
-                    $"<color=red>{nameof(muteKey)}</color>: <color=blue>{mute}</color>");
+                bool mute = (PlayerPrefs.HasKey(muteKey) && PlayerPrefs.GetInt(muteKey) == 1);
+                GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
+                        $"<color=red>{nameof(muteKey)}</color>: <color=blue>{mute}</color>");
 
-            GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
-                    $"<color=red>{nameof(AudioPlayer.VolumeKey)}</color>: <color=blue>{AudioPlayer.GlobalVolume}</color>");
+                GUI.Label(new Rect(new Vector2(10, CurrentOffset), rectSize),
+                        $"<color=red>{nameof(AudioPlayer.VolumeKey)}</color>: <color=blue>{AudioPlayer.GlobalVolume}</color>");
 
-            if (GUI.Button(new Rect(new Vector2(10, CurrentOffset), rectSize), "Delete All Keys")) {
-                PlayerPrefs.DeleteAll();
+                if (GUI.Button(new Rect(new Vector2(10, CurrentOffset), rectSize), "Delete All Keys")) {
+                    PlayerPrefs.DeleteAll();
+                }
             }
         }
